refactor: extract http poll mixin from coordinator core
